requirements.txt
pytest
selenium
pytest-html
webdriver-manager
pytest-reportlog
allure-pytest